backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
psycopg2-binary
python-dotenv
httpx
google-genai
pydantic
pydantic-settings
python-jose
passlib